+++ examples/advanced.rs
@@ -2,8 +2,10 @@ use argone::{prelude::*, *};
 use std::path::PathBuf;
 
 ARGS! {
-    // this area is top of the clap struct meta area.
+    /// this area is top of the clap struct meta area.
+    #[derive(Clone)]
     #[clap(after_help = "My Application")]
+
     version = "0.1"
     author = "just-do-halee <just.do.halee@gmail.com>"
 
@@ -15,21 +17,27 @@ ARGS! {
 
     Args {
         /// Root directory
-        [Config] rootDir: Option<PathBuf> = CURRENT_DIR.clone()
+        [Config] rootDir: Option<PathBuf> = Some(CURRENT_DIR.clone())
 
         /// This name will be changed as app-name or APP_NAME
-        #[clap(short, long, default_value = "test")]
-        appName: String
+        (short, long, default_value = "test")
+        appName: String; // can be separated by ';' or ','
+
+        /// Some List
+        (short, long)
+        [Config] someList: Vec<String> = vec!["default".to_string()]
 
         /// A level of verbosity, and can be used multiple times
-        #[clap(short, long, parse(from_occurrences))]
-        verbose: u8
+        (short, long, parse(from_occurrences))
+        verbose: u8,
     }
 
     commands = Sub2
 }
 
 COMMANDS! {
+    /// this area is top of the command enum meta area.
+    #[derive(Copy)]
     Sub {
         /// third
         Third {
@@ -39,9 +47,11 @@ COMMANDS! {
             }
         }
     }
+
+    /// same
     Sub2 {
         /// first
-        #[clap(alias = "foo")]
+        (alias = "foo")
         First {
             version = "1.0"
             author = "just-do-halee <just.do.halee@gmail.com>"
@@ -51,7 +61,7 @@ COMMANDS! {
             }
         }
         /// second
-        #[clap(aliases = &["baz", "fizz"])]
+        (aliases = &["baz", "fizz"])
         Second {
             Args {
                 // Test u8