+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 
+## 0.4.0 (December 3, 2021)
+
+### Release 0.4.0
+
+* removing unnecessary #[clap()] tokens.
+* now [Config] can parse Vector types.
+* fixing small bugs.
+
+[Example]
+* (short, long)
+* [Config] someList: Vec\<String\> = vec!["default".to_string()]
+
+---
+
 ## 0.3.0 (December 2, 2021)
 
 ### Release 0.3.0